Make ServerBuild the one reader of the build stamp
